+++ NEWS.md
@@ -30,6 +30,8 @@
 
 - `formatDate()` may display dates off by one day (thanks, @jdthorpe, #198).
 
+- When a table is destroyed, all Shiny input values associated with the table should be emptied (thanks, @dracodoc, #457).
+
 # CHANGES IN DT VERSION 0.2
 
 ## NEW FEATURES